refactor: Extract common lambda code in shared module
It's included in every zip file and every lambda module that imports it should import it as "from common import *".
